# docs/openapi-v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "乒乓球赛事编排与赛务管理系统 Demo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tournaments/{tournament_id}": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "summary": "Get Tournament",
        "operationId": "get_tournament_api_tournaments__tournament_id__get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tournaments"
        ],
        "summary": "Delete Tournament",
        "description": "删除赛事（级联删除分组/选手/球台/比赛，见 db.py 外键 ON DELETE CASCADE）。",
        "operationId": "delete_tournament_api_tournaments__tournament_id__delete",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/players/import": {
      "post": {
        "tags": [
          "players"
        ],
        "summary": "Import Players",
        "operationId": "import_players_api_tournaments__tournament_id__players_import_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_players_api_tournaments__tournament_id__players_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPlayersResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/players/import/preview": {
      "post": {
        "tags": [
          "players"
        ],
        "summary": "Preview Import Players",
        "operationId": "preview_import_players_api_tournaments__tournament_id__players_import_preview_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_import_players_api_tournaments__tournament_id__players_import_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/players": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "List Players",
        "operationId": "list_players_api_tournaments__tournament_id__players_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerOut"
                  },
                  "title": "Response List Players Api Tournaments  Tournament Id  Players Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "players"
        ],
        "summary": "Add Player",
        "operationId": "add_player_api_tournaments__tournament_id__players_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/players/{player_id}": {
      "patch": {
        "tags": [
          "players"
        ],
        "summary": "Update Player",
        "operationId": "update_player_api_tournaments__tournament_id__players__player_id__patch",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          },
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "players"
        ],
        "summary": "Delete Player",
        "operationId": "delete_player_api_tournaments__tournament_id__players__player_id__delete",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          },
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments/{tournament_id}/generate-group-matches": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Generate Group Matches",
        "operationId": "generate_group_matches_api_tournaments__tournament_id__generate_group_matches_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateMatchesResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "List Matches",
        "operationId": "list_matches_api_tournaments__tournament_id__matches_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "GROUP",
                    "KNOCKOUT"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "WAITING",
                    "PLAYING",
                    "FINISHED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchOut"
                  },
                  "title": "Response List Matches Api Tournaments  Tournament Id  Matches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_players_api_tournaments__tournament_id__players_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_players_api_tournaments__tournament_id__players_import_post"
      },
      "Body_preview_import_players_api_tournaments__tournament_id__players_import_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_import_players_api_tournaments__tournament_id__players_import_preview_post"
      },
      "BronzeMode": {
        "type": "string",
        "enum": [
          "BRONZE_MATCH",
          "JOINT_BRONZE"
        ],
        "title": "BronzeMode"
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "SINGLES",
          "DOUBLES"
        ],
        "title": "EventType",
        "description": "参赛项目。第一版完整支持单打与固定搭档双打。"
      },
...
      "GenerateMatchesResult": {
        "properties": {
          "matches_generated": {
            "type": "integer",
            "title": "Matches Generated"
          },
          "per_group": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Per Group"
          },
          "tournament": {
            "$ref": "#/components/schemas/TournamentOut"
          }
        },
        "type": "object",
        "required": [
          "matches_generated",
          "per_group",
          "tournament"
        ],
        "title": "GenerateMatchesResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportPlayersResult": {
        "properties": {
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "imported": {
            "type": "integer",
            "title": "Imported"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportRowError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "total_rows",
          "imported",
          "skipped",
          "errors"
        ],
        "title": "ImportPlayersResult"
      },
      "ImportPreviewResult": {
        "properties": {
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "valid_rows": {
            "type": "integer",
            "title": "Valid Rows"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportRowError"
            },
            "type": "array",
            "title": "Errors"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportPreviewRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total_rows",
          "valid_rows",
          "skipped",
          "errors",
          "rows"
        ],
        "title": "ImportPreviewResult"
      },
...
      "MatchOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "stage": {
            "$ref": "#/components/schemas/MatchStage"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "match_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Index"
          },
          "player_a_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player A Id"
          },
          "player_b_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player B Id"
          },
          "player_a_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player A Score"
          },
          "player_b_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player B Score"
          },
          "winner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Id"
          },
          "table_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Id"
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "prev_match_a_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Match A Id"
          },
          "prev_match_b_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Match B Id"
          },
          "entry_a_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry A Id"
          },
          "entry_b_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry B Id"
          },
          "winner_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Entry Id"
          },
          "entry_a_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry A Name"
          },
          "entry_b_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry B Name"
          },
          "result_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultType"
              },
              {
                "type": "null"
              }
            ]
          },
          "forfeit_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forfeit Entry Id"
          },
          "bracket": {
            "$ref": "#/components/schemas/MatchBracket",
            "default": "GROUP"
          },
          "placement_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement Min"
          },
          "placement_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement Max"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/MatchGameOut"
            },
            "type": "array",
            "title": "Games",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournament_id",
          "stage",
          "group_id",
          "round",
          "match_index",
          "player_a_id",
          "player_b_id",
          "player_a_score",
          "player_b_score",
          "winner_id",
          "table_id",
          "status",
          "prev_match_a_id",
          "prev_match_b_id"
        ],
        "title": "MatchOut"
      },
      "MatchStage": {
        "type": "string",
        "enum": [
          "GROUP",
          "KNOCKOUT"
        ],
        "title": "MatchStage",
        "description": "比赛所属赛段。"
      },
...
      "PlacementMode": {
        "type": "string",
        "enum": [
          "OFF",
          "COMPLETE",
          "TIERED"
        ],
        "title": "PlacementMode"
      },
...
      "PlayerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "college": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "College"
          },
          "rating_points": {
            "type": "integer",
            "maximum": 99999.0,
            "minimum": 0.0,
            "title": "Rating Points",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PlayerCreate"
      },
      "PlayerOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "college": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "seed_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed No"
          },
          "rating_points": {
            "type": "integer",
            "title": "Rating Points",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournament_id",
          "name",
          "college",
          "group_id",
          "seed_no"
        ],
        "title": "PlayerOut"
      },
      "PlayerUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "college": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "College"
          },
          "rating_points": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 99999.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Points"
          }
        },
        "type": "object",
        "title": "PlayerUpdate"
      },
...
      "TournamentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "table_count": {
            "type": "integer",
            "title": "Table Count"
          },
          "group_count": {
            "type": "integer",
            "title": "Group Count"
          },
          "qualify_per_group": {
            "type": "integer",
            "title": "Qualify Per Group"
          },
          "stage": {
            "$ref": "#/components/schemas/TournamentStage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType",
            "default": "SINGLES"
          },
          "bronze_mode": {
            "$ref": "#/components/schemas/BronzeMode",
            "default": "JOINT_BRONZE"
          },
          "placement_mode": {
            "$ref": "#/components/schemas/PlacementMode",
            "default": "OFF"
          },
          "games_to_win": {
            "type": "integer",
            "title": "Games To Win",
            "default": 2
          },
          "points_to_win": {
            "type": "integer",
            "title": "Points To Win",
            "default": 11
          },
          "roster_confirmed": {
            "type": "boolean",
            "title": "Roster Confirmed",
            "default": false
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "date",
          "table_count",
          "group_count",
          "qualify_per_group",
          "stage",
          "created_at"
        ],
        "title": "TournamentOut"
      },
      "TournamentStage": {
        "type": "string",
        "enum": [
          "REGISTRATION",
          "GROUP_STAGE",
          "KNOCKOUT",
          "FINISHED"
        ],
        "title": "TournamentStage",
        "description": "赛事阶段。"
      },
...
    }
  }
}